Move MediaQueryFromWindow to media_query
